Change implementation of trans_cnot
