+++ tests/log_gates/test_unrot_surface_code_css.py
@@ -1,7 +1,6 @@
 from surface_sim.log_gates.unrot_surface_code_css import (
     set_fold_trans_s,
     set_fold_trans_h,
-    set_trans_cnot,
 )
 from surface_sim.layouts import unrot_surface_code
 from surface_sim.detectors import get_new_stab_dict_from_layout
@@ -72,65 +71,6 @@ def test_set_fold_trans_s():
         assert len(new_stabs[z]) == 1
     for x in x_stabs:
         assert len(new_stabs[x]) == 2
-
-    return
-
-
-def test_set_trans_cnot():
-    distance = 9
-    layout_c = unrot_surface_code(distance=distance)
-    layout_t = unrot_surface_code(
-        distance=distance,
-        logical_qubit_label="L1",
-        init_point=(1000, 1000),
-        init_data_qubit_id=1000,
-        init_zanc_qubit_id=1000,
-        init_xanc_qubit_id=1000,
-        init_ind=layout_c.get_max_ind() + 1,
-    )
-    set_trans_cnot(layout_c, layout_t)
-    gate_label = f"log_trans_cnot_{layout_c.get_logical_qubits()[0]}_{layout_t.get_logical_qubits()[0]}"
-
-    x_stabs = layout_c.get_qubits(role="anc", stab_type="x_type")
-    for key in ["new_stab_gen", "new_stab_gen_inv"]:
-        new_stab_x = [layout_c.param(gate_label, x_stab)[key] for x_stab in x_stabs]
-        assert new_stab_x == [[i, f"X{int(i[1:])+1000-1}"] for i in x_stabs]
-
-    x_stabs = layout_t.get_qubits(role="anc", stab_type="x_type")
-    for key in ["new_stab_gen", "new_stab_gen_inv"]:
-        new_stab_x = [layout_t.param(gate_label, x_stab)[key] for x_stab in x_stabs]
-        assert new_stab_x == [[i] for i in x_stabs]
-
-    z_stabs = layout_t.get_qubits(role="anc", stab_type="z_type")
-    for key in ["new_stab_gen", "new_stab_gen_inv"]:
-        new_stab_z = [layout_t.param(gate_label, z_stab)[key] for z_stab in z_stabs]
-        assert new_stab_z == [[i, f"Z{int(i[1:])-1000+1}"] for i in z_stabs]
-
-    z_stabs = layout_c.get_qubits(role="anc", stab_type="z_type")
-    for key in ["new_stab_gen", "new_stab_gen_inv"]:
-        new_stab_z = [layout_c.param(gate_label, z_stab)[key] for z_stab in z_stabs]
-        assert new_stab_z == [[i] for i in z_stabs]
-
-    data_qubits = layout_c.get_qubits(role="data")
-    cz_gates = [layout_c.param(gate_label, d)["cnot"] for d in data_qubits]
-    assert cz_gates == [f"D{int(i[1:])+1000-1}" for i in data_qubits]
-
-    new_stabs, new_stabs_inv = get_new_stab_dict_from_layout(layout_c, gate_label)
-    assert new_stabs == new_stabs_inv
-    x_stabs = layout_c.get_qubits(role="anc", stab_type="x_type")
-    z_stabs = layout_c.get_qubits(role="anc", stab_type="z_type")
-    for z in z_stabs:
-        assert len(new_stabs[z]) == 1
-    for x in x_stabs:
-        assert len(new_stabs[x]) == 2
-    new_stabs, new_stabs_inv = get_new_stab_dict_from_layout(layout_t, gate_label)
-    assert new_stabs == new_stabs_inv
-    x_stabs = layout_t.get_qubits(role="anc", stab_type="x_type")
-    z_stabs = layout_t.get_qubits(role="anc", stab_type="z_type")
-    for z in z_stabs:
-        assert len(new_stabs[z]) == 2
-    for x in x_stabs:
-        assert len(new_stabs[x]) == 1
 
     return
